requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy[asyncio]
alembic
asyncpg
pydantic
pydantic-settings
boto3
deepgram-sdk
anthropic
sendgrid
celery[redis]
redis
python-multipart
streamlit
python-dotenv
httpx
tenacity
psycopg2-binary
PyJWT
bcrypt
email-validator
